Noticed another _apply that should be eval_
