+++ mathics/builtin/messages.py
@@ -230,6 +230,7 @@ class General(Builtin):
         "ssym": "`1` is not a symbol or a string.",
         "stream": "`1` is not string, InputStream[], or OutputStream[]",
         "string": "String expected.",
+        "strse": "String or list of strings expected at position `1` in `2`.",
         "sym": "Argument `1` at position `2` is expected to be a symbol.",
         "tag": "Rule for `1` can only be attached to `2`.",
         "take": "Cannot take positions `1` through `2` in `3`.",
