+++ mathics/eval/strings.py
@@ -1,11 +1,17 @@
 """
 String-related evaluation functions.
 """
+import re
 
-from mathics.core.atoms import String
+from mathics.core.atoms import Integer1, Integer3, String
+from mathics.core.convert.expression import to_mathics_list
+from mathics.core.convert.regex import to_regex
 from mathics.core.element import BaseElement
 from mathics.core.evaluation import Evaluation
-from mathics.core.symbols import Symbol
+from mathics.core.expression import Expression
+from mathics.core.expression_predefined import MATHICS3_INFINITY
+from mathics.core.list import ListExpression
+from mathics.core.symbols import Symbol, SymbolTrue
 from mathics.eval.makeboxes import format_element
 
 
@@ -16,3 +22,77 @@ def eval_ToString(
     boxes = format_element(expr, evaluation, form, encoding=encoding)
     text = boxes.boxes_to_text(evaluation=evaluation)
     return String(text)
+
+
+def eval_StringFind(self, string, rule, n, evaluation, options, cases):
+    if n.sameQ(Symbol("System`Private`Null")):
+        expr = Expression(Symbol(self.get_name()), string, rule)
+        n = None
+    else:
+        expr = Expression(Symbol(self.get_name()), string, rule, n)
+
+    # convert string
+    if isinstance(string, ListExpression):
+        py_strings = [stri.get_string_value() for stri in string.elements]
+        if None in py_strings:
+            evaluation.message(self.get_name(), "strse", Integer1, expr)
+            return
+    else:
+        py_strings = string.get_string_value()
+        if py_strings is None:
+            evaluation.message(self.get_name(), "strse", Integer1, expr)
+            return
+
+    # convert rule
+    def convert_rule(r):
+        if r.has_form("Rule", None) and len(r.elements) == 2:
+            py_s = to_regex(r.elements[0], show_message=evaluation.message)
+            if py_s is None:
+                evaluation.message(
+                    "StringExpression", "invld", r.elements[0], r.elements[0]
+                )
+                return
+            py_sp = r.elements[1]
+            return py_s, py_sp
+        elif cases:
+            py_s = to_regex(r, show_message=evaluation.message)
+            if py_s is None:
+                evaluation.message("StringExpression", "invld", r, r)
+                return
+            return py_s, None
+
+        evaluation.message(self.get_name(), "srep", r)
+        return
+
+    if rule.has_form("List", None):
+        py_rules = [convert_rule(r) for r in rule.elements]
+    else:
+        py_rules = [convert_rule(rule)]
+    if None in py_rules:
+        return None
+
+    # convert n
+    if n is None:
+        py_n = 0
+    elif n.sameQ(MATHICS3_INFINITY):
+        py_n = 0
+    else:
+        py_n = n.get_int_value()
+        if py_n is None or py_n < 0:
+            evaluation.message(self.get_name(), "innf", Integer3, expr)
+            return
+
+    # flags
+    flags = re.MULTILINE
+    if options["System`IgnoreCase"] is SymbolTrue:
+        flags = flags | re.IGNORECASE
+
+    if isinstance(py_strings, list):
+        return to_mathics_list(
+            *[
+                self._find(py_stri, py_rules, py_n, flags, evaluation)
+                for py_stri in py_strings
+            ]
+        )
+    else:
+        return self._find(py_strings, py_rules, py_n, flags, evaluation)
